fixes #866 move TlsUtil to config module from utility to remove config dependency
