+++ client/src/main/java/com/networknt/client/Http2Client.java
@@ -27,7 +27,6 @@ import com.networknt.client.oauth.TokenManager;
 import com.networknt.client.ssl.ClientX509ExtendedTrustManager;
 import com.networknt.client.ssl.TLSConfig;
 import com.networknt.cluster.Cluster;
-import com.networknt.common.SecretConstants;
 import com.networknt.config.Config;
 import com.networknt.httpstring.AttachmentConstants;
 import com.networknt.exception.ClientException;
@@ -37,7 +36,7 @@ import com.networknt.monad.Result;
 import com.networknt.service.SingletonServiceFactory;
 import com.networknt.status.Status;
 import com.networknt.utility.ModuleRegistry;
-import com.networknt.utility.TlsUtil;
+import com.networknt.config.TlsUtil;
 import io.opentracing.Tracer;
 import io.opentracing.propagation.Format;
 import io.opentracing.tag.Tags;
