+++ config/src/test/java/com/networknt/config/TlsUtilTest.java
@@ -1,5 +1,6 @@
-package com.networknt.utility;
+package com.networknt.config;
 
+import com.networknt.config.TlsUtil;
 import org.junit.Assert;
 import org.junit.Test;
