+++ server/src/main/java/com/networknt/server/Server.java
@@ -30,8 +30,7 @@ import com.networknt.status.Status;
 import com.networknt.switcher.SwitcherUtil;
 import com.networknt.utility.Constants;
 import com.networknt.utility.NetUtils;
-import com.networknt.utility.TlsUtil;
-import com.networknt.utility.Util;
+import com.networknt.config.TlsUtil;
 import io.undertow.Handlers;
 import io.undertow.Undertow;
 import io.undertow.UndertowOptions;
